--- src/type.ts
@@ -1,9 +1,0 @@
-export type CardData = {
-  id: number;
-  location: string;
-  name: string;
-  url: string;
-  capacity: string;
-  year: number;
-  used: string;
-};
